refactor: ブートオーバーレイをモジュール化して呼び出しを簡素化
